Split external-review miss test suites
