+++ .codex-supervisor/issue-journal.md
@@ -1,44 +1,35 @@
-# Issue #457: CodeRabbit settled wait: add regression coverage for CI-green immediate-merge races
+# Issue #462: Test refactor: split external-review miss tests by normalization and durable-guardrail coverage
 
 ## Supervisor Snapshot
-- Issue URL: https://github.com/TommyKammy/codex-supervisor/issues/457
-- Branch: codex/issue-457
-- Workspace: .
-- Journal: .codex-supervisor/issue-journal.md
-- Current phase: waiting_ci
-- Attempt count: 3 (implementation=1, repair=2)
-- Last head SHA: 119e18e316763d97d9c8e22ab62b8c55e22114a7
+- Issue URL: https://github.com/TommyKammy/codex-supervisor/issues/462
+- Branch: codex/issue-462
+- Workspace: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-462
+- Journal: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-462/.codex-supervisor/issue-journal.md
+- Current phase: reproducing
+- Attempt count: 1 (implementation=1, repair=0)
+- Last head SHA: 9f566dee933873fd345c89c4a3a7a0fea426ed9d
 - Blocked reason: none
 - Last failure signature: none
 - Repeated failure signature count: 0
-- Updated at: 2026-03-17T07:33:05.815Z
+- Updated at: 2026-03-17T08:27:30Z
 
 ## Latest Codex Summary
-Addressed the remaining CodeRabbit review finding on PR #460 by sanitizing the committed journal metadata, committed the repair as `119e18e`, pushed `codex/issue-457`, and resolved the CodeRabbit thread via GitHub.
-
-No code or test fixtures changed in this turn. The PR is no longer draft, the review thread is resolved, and CI restarted on the updated branch; at the time of this note, `build (macos-latest)` had already passed while `build (ubuntu-latest)` and `CodeRabbit` were still pending.
-
-Summary: Pushed the journal-only review fix for PR #460, resolved the CodeRabbit thread, and handed the branch off to CI.
-State hint: waiting_ci
-Blocked reason: none
-Tests: not run; journal-only review fix
-Failure signature: none
-Next action: Monitor the refreshed PR #460 checks and only intervene if CI or follow-up review reports a new issue.
+- Split the monolithic external-review misses suite into focused normalization, classifier, persistence, and miss-history test files; removed the old combined test file after preserving its assertions.
 
 ## Active Failure Context
 - None recorded.
 
 ## Codex Working Notes
 ### Current Handoff
-- Hypothesis: CodeRabbit can emit later review-thread comments with missing `originalCommitOid`; those comments are still safe to treat as current-head activity only when a stronger CodeRabbit current-head observation already exists.
-- What changed: Added focused review-signal and hydrator tests for weakly anchored CodeRabbit review comments, including a stale-head negative case; updated `inferConfiguredBotCurrentHeadObservedAt` to extend `currentHeadObservedAt` only for later CodeRabbit comments lacking `originalCommitOid` after a prior strong current-head observation. This turn also sanitized the issue journal metadata to remove absolute local paths called out in review.
+- Hypothesis: The external-review miss coverage is easier to maintain when assertions live beside the normalization, classifier, persistence, and miss-history module seams instead of a single shared suite.
+- What changed: Added `external-review-normalization.test.ts`, `external-review-classifier.test.ts`, `external-review-miss-persistence.test.ts`, and `external-review-miss-history.test.ts`; moved the existing assertions into those files with local fixtures and deleted `external-review-misses.test.ts`.
 - Current blocker: none
-- Next exact step: Wait for the refreshed PR #460 checks to finish and react only if CI or review opens a new failure.
-- Verification gap: none for the focused review-signal, pull-request-state, supervisor lifecycle, hydrator, and build checks requested by the issue; no additional verification is needed for this journal-only repair.
-- Files touched: src/github/github-review-signals.ts; src/github/github-review-signals.test.ts; src/github/github-pull-request-hydrator.test.ts; .codex-supervisor/issue-journal.md
-- Rollback concern: The weak-anchor extension is intentionally limited to CodeRabbit-authored review-thread comments that happen after a strong current-head observation; broadening that rule to other bots or comments without the prior observation would risk promoting stale history.
-- Last focused command: gh pr view 460 --json isDraft,mergeStateStatus,reviewDecision,statusCheckRollup,headRefName
+- Next exact step: Commit the extracted test suites on `codex/issue-462` and open or update the draft PR if needed.
+- Verification gap: none for the requested external-review normalization, misses, guardrail-loading, and build checks after `npm ci` restored missing local dependencies.
+- Files touched: src/external-review/external-review-normalization.test.ts; src/external-review/external-review-classifier.test.ts; src/external-review/external-review-miss-persistence.test.ts; src/external-review/external-review-miss-history.test.ts; src/external-review/external-review-misses.test.ts; .codex-supervisor/issue-journal.md
+- Rollback concern: The new suites intentionally duplicate a few tiny fixture builders to keep module-local coverage independent; collapsing them back into a shared helper would reintroduce cross-file coupling and blur the boundaries this issue is trying to restore.
+- Last focused command: npm run build
 ### Scratchpad
 - Keep this section short. The supervisor may compact older notes automatically.
-- Reproducing signature before the fix: `currentHeadObservedAt` stayed at the earlier summary-only current-head review and did not advance to a later CodeRabbit review-thread comment with `originalCommitOid: null`.
-- Verification commands: `npx tsx --test src/github/github-review-signals.test.ts`; `npx tsx --test src/pull-request-state.test.ts`; `npx tsx --test src/supervisor/supervisor-lifecycle.test.ts`; `npm ci`; `npx tsx --test src/github/github-pull-request-hydrator.test.ts`; `npm run build`; `rg -n '^- (Workspace|Journal):' .codex-supervisor/issue-journal.md`; `git commit -m "Sanitize issue journal paths"`; `git push origin codex/issue-457`; `gh api graphql -f query='mutation($threadId: ID!) { resolveReviewThread(input: {threadId: $threadId}) { thread { id isResolved } } }' -F threadId=PRRT_kwDORgvdZ850x8eI`; `gh pr view 460 --json isDraft,mergeStateStatus,reviewDecision,statusCheckRollup,headRefName`.
+- Reproducing signature before the fix: `external-review-misses.test.ts` mixed normalization, matching, persistence, and durable guardrail loading in one 22-test file, so edits and failures crossed module boundaries.
+- Verification commands: `npx tsx --test src/external-review/external-review-normalization.test.ts`; `npx tsx --test src/external-review/external-review-classifier.test.ts`; `npx tsx --test src/external-review/external-review-miss-persistence.test.ts`; `npx tsx --test src/external-review/external-review-miss-history.test.ts`; `npx tsx --test src/external-review/*.test.ts`; `npm ci`; `npm run build`.
